microservices/ayurbot-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.4.1
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
certifi==2024.2.2
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
tiktoken
